Fix code factor warnings 🚨
