+++ Src/Notion.Client/Models/Blocks/ColumnListBlock.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections.Generic;
+using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    public class ColumnListBlock : Block, INonColumnBlock
+    {
+        public override BlockType Type => BlockType.ColumnList;
+
+        [JsonProperty("column_list")]
+        public Info ColumnList { get; set; }
+
+        public class Info
+        {
+            [JsonProperty("children")]
+            public IEnumerable<ColumnBlock> Children { get; set; }
+        }
+    }
+}
